+++ .github/workflows/helpers/auth0/.gitignore
@@ -1,0 +1,2 @@
+/node_modules
+/local
