+++ .github/workflows/helpers/auth0/tenant.yaml
@@ -1,0 +1,259 @@
+pages: []
+resourceServers:
+  - name: paradedb-api
+    identifier: https://provision-dev.cloud.paradedb.com
+    allow_offline_access: false
+    signing_alg: RS256
+    skip_consent_for_verifiable_first_party_clients: true
+    token_lifetime: 86400
+    token_lifetime_for_web: 7200
+clients:
+  - name: ParadeDB Dashboard
+    allowed_clients: []
+    allowed_logout_urls:
+      - http://localhost:3000
+      - https://paradedb.vercel.app
+    app_type: regular_web
+    callbacks:
+      - http://localhost:3000/api/auth/callback
+      - https://paradedb.vercel.app/api/auth/callback
+    client_aliases: []
+    cross_origin_auth: false
+    custom_login_page_on: true
+    grant_types:
+      - authorization_code
+      - implicit
+      - refresh_token
+      - client_credentials
+    is_first_party: true
+    is_token_endpoint_ip_header_trusted: false
+    jwt_configuration:
+      alg: RS256
+      lifetime_in_seconds: 36000
+      secret_encoded: false
+    logo_uri: https://raw.githubusercontent.com/paradedb/paradedb/dev/docs/images/icon_logo.png
+    native_social_login:
+      apple:
+        enabled: false
+      facebook:
+        enabled: false
+    oidc_conformant: true
+    refresh_token:
+      expiration_type: non-expiring
+      leeway: 0
+      infinite_token_lifetime: true
+      infinite_idle_token_lifetime: true
+      token_lifetime: 31557600
+      idle_token_lifetime: 2592000
+      rotation_type: non-rotating
+    sso_disabled: false
+    token_endpoint_auth_method: client_secret_post
+  - name: paradedb-api (Test Application)
+    app_type: non_interactive
+    cross_origin_auth: false
+    custom_login_page_on: true
+    grant_types:
+      - client_credentials
+    is_first_party: true
+    is_token_endpoint_ip_header_trusted: false
+    jwt_configuration:
+      alg: RS256
+      lifetime_in_seconds: 36000
+      secret_encoded: false
+    oidc_conformant: true
+    refresh_token:
+      expiration_type: non-expiring
+      leeway: 0
+      infinite_token_lifetime: true
+      infinite_idle_token_lifetime: true
+      token_lifetime: 31557600
+      idle_token_lifetime: 2592000
+      rotation_type: non-rotating
+    sso_disabled: false
+    token_endpoint_auth_method: client_secret_post
+databases:
+  - name: Username-Password-Authentication
+    strategy: auth0
+    enabled_clients:
+      - ParadeDB Dashboard
+      - paradedb-api (Test Application)
+    is_domain_connection: false
+    options:
+      mfa:
+        active: true
+        return_enroll_settings: true
+      passwordPolicy: good
+      passkey_options:
+        challenge_ui: both
+        local_enrollment_enabled: true
+        progressive_enrollment_enabled: true
+      strategy_version: 2
+      authentication_methods:
+        passkey:
+          enabled: false
+        password:
+          enabled: true
+      brute_force_protection: true
+    realms:
+      - Username-Password-Authentication
+connections:
+  - name: google-oauth2
+    strategy: google-oauth2
+    enabled_clients:
+      - ParadeDB Dashboard
+      - paradedb-api (Test Application)
+    is_domain_connection: false
+    options:
+      email: true
+      scope:
+        - email
+        - profile
+      profile: true
+tenant:
+  enabled_locales:
+    - en
+    - eu-ES
+    - bs
+    - bg
+    - ca-ES
+    - zh-CN
+    - zh-TW
+    - hr
+    - cs
+    - da
+    - nl
+    - et
+    - fi
+    - fr-CA
+    - fr
+    - gl-ES
+    - de
+    - el
+    - hi
+    - hu
+    - is
+    - id
+    - it
+    - ja
+    - ko
+    - lv
+    - lt
+    - "no"
+    - nb
+    - nn
+    - pl
+    - pt
+    - pt-PT
+    - ro
+    - ru
+    - sr
+    - sk
+    - sl
+    - es
+    - sv
+    - th
+    - tr
+    - uk
+    - vi
+    - cy
+  flags:
+    revoke_refresh_token_grant: false
+    disable_clickjack_protection_headers: false
+  friendly_name: ParadeDB
+  picture_url: https://raw.githubusercontent.com/paradedb/paradedb/dev/docs/images/icon_logo.png
+  sandbox_version: "18"
+  support_email: support@paradedb.com
+  support_url: https://paradedb.com
+  universal_login:
+    colors:
+      page_background: "#efefef"
+      primary: "#34d399"
+emailProvider: {}
+emailTemplates: []
+clientGrants:
+  - client_id: ParadeDB Dashboard
+    audience: https://provision-dev.cloud.paradedb.com
+    scope: []
+guardianFactors:
+  - name: duo
+    enabled: false
+  - name: email
+    enabled: false
+  - name: otp
+    enabled: false
+  - name: push-notification
+    enabled: false
+  - name: recovery-code
+    enabled: false
+  - name: sms
+    enabled: false
+  - name: webauthn-platform
+    enabled: false
+  - name: webauthn-roaming
+    enabled: false
+guardianFactorProviders: []
+guardianFactorTemplates: []
+guardianPolicies:
+  policies: []
+guardianPhoneFactorSelectedProvider:
+  provider: auth0
+guardianPhoneFactorMessageTypes:
+  message_types: []
+roles: []
+branding:
+  colors:
+    page_background: "#efefef"
+    primary: "#34d399"
+  logo_url: https://raw.githubusercontent.com/paradedb/paradedb/dev/docs/images/icon_logo.png
+  templates: []
+prompts:
+  customText: {}
+  identifier_first: false
+  universal_login_experience: new
+migrations: {}
+actions:
+  - name: create-stripe-customer
+    code: ./actions/create-stripe-customer/code.js
+    dependencies:
+      - name: stripe
+        version: 14.5.0
+    deployed: true
+    secrets:
+      - name: STRIPE_API_KEY
+    supported_triggers:
+      - id: post-login
+        version: v2
+triggers: {}
+organizations: []
+attackProtection:
+  breachedPasswordDetection:
+    enabled: false
+    shields: []
+    admin_notification_frequency: []
+    method: standard
+    stage:
+      pre-user-registration:
+        shields: []
+  bruteForceProtection:
+    enabled: true
+    shields:
+      - block
+      - user_notification
+    mode: count_per_identifier_and_ip
+    allowlist: []
+    max_attempts: 10
+  suspiciousIpThrottling:
+    enabled: true
+    shields:
+      - admin_notification
+      - block
+    allowlist: []
+    stage:
+      pre-login:
+        max_attempts: 100
+        rate: 864000
+      pre-user-registration:
+        max_attempts: 50
+        rate: 1200
+logStreams: []
+themes: []
