+++ dashboard/app/api/databases/credentials/route.ts
@@ -1,6 +1,6 @@
 import { withRequest } from "@/utils/api";
 
-const GET = withRequest((accessToken: string) =>
+const GET = withRequest(({ accessToken }) =>
   fetch(`${process.env.PROVISIONER_URL}/databases/credentials`, {
     method: "GET",
     headers: {
