+++ dashboard/components/modal/index.ts
@@ -1,0 +1,2 @@
+export * from "./configure-instance";
+export * from "./payment-intent";
